+++ js/packages/openinference-instrumentation-openai/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @arizeai/openinference-instrumentation-openai
 
+## 2.0.2
+
+### Patch Changes
+
+- f1c1a3c: place image attributes under the right path
+
+## 2.0.1
+
+### Patch Changes
+
+- b69cca6: Account for developer role in the openai instrumentation
+
 ## 2.0.0
 
 ### Major Changes