+++ python/instrumentation/openinference-instrumentation-llama-index/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [3.3.3](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-llama-index-v3.3.2...python-openinference-instrumentation-llama-index-v3.3.3) (2025-03-18)
+
+
+### Bug Fixes
+
+* replace logger.exception() with repr() ([#1385](https://github.com/Arize-ai/openinference/issues/1385)) ([3b924da](https://github.com/Arize-ai/openinference/commit/3b924da0b823e23c9170a6dae682bb0488c884cb))
+
+## [3.3.2](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-llama-index-v3.3.1...python-openinference-instrumentation-llama-index-v3.3.2) (2025-03-14)
+
+
+### Documentation
+
+* fix license to be openinference ([#1353](https://github.com/Arize-ai/openinference/issues/1353)) ([85d435b](https://github.com/Arize-ai/openinference/commit/85d435be3af3de5424494cfbdd654454688b7377))
+
 ## [3.3.1](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-llama-index-v3.3.0...python-openinference-instrumentation-llama-index-v3.3.1) (2025-02-19)